docs : authController Profile 별 API 분리 및 Swagger 프로필 설정
