+++ src/main/java/Gdsc/web/config/corsConfig/CorsConfigReal.java
@@ -19,7 +19,6 @@ public class CorsConfigReal implements CorsConfig {
         config.setAllowCredentials(true); //  내 서버가 응답할때 json을 자바스크립트가 처리할수있게 할건지
 
         config.addAllowedOriginPattern("https://gdsc-dju-blog.web.app");
-        config.addAllowedOriginPattern("https://gdsc-dju-blog.web.com");
         config.addAllowedHeader("*"); // 모든 헤더 응답허용
         config.addAllowedMethod("*"); // 모든 post get put delete 요청 허용
         source.registerCorsConfiguration("/**", config);
