+++ src/main/java/Gdsc/web/controller/auth/RefleshController.java
@@ -1,0 +1,100 @@
+package Gdsc.web.controller.auth;
+
+import Gdsc.web.config.properties.AppProperties;
+import Gdsc.web.dto.ApiResponse;
+import Gdsc.web.entity.UserRefreshToken;
+import Gdsc.web.model.RoleType;
+import Gdsc.web.oauth.token.AuthToken;
+import Gdsc.web.oauth.token.AuthTokenProvider;
+import Gdsc.web.repository.UserRefreshTokenRepository;
+import Gdsc.web.service.MemberService;
+import Gdsc.web.utils.CookieUtil;
+import Gdsc.web.utils.HeaderUtil;
+import io.jsonwebtoken.Claims;
+import io.swagger.annotations.ApiOperation;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.Date;
+@RestController
+@RequestMapping("")
+@RequiredArgsConstructor
+public class RefleshController {
+    private final AppProperties appProperties;
+    private final AuthTokenProvider tokenProvider;
+    private final AuthenticationManager authenticationManager;
+    private final UserRefreshTokenRepository userRefreshTokenRepository;
+    private  final MemberService memberService;
+    private final static long THREE_DAYS_MSEC = 259200000;
+    private final static String REFRESH_TOKEN = "refresh_token";
+    @GetMapping("/refresh")
+    @ApiOperation(value = "refresh 토큰을 이용하여 JWT 토큰 재발급", notes = "토큰이 expired 되어야 작동함")
+    public ApiResponse refreshToken (HttpServletRequest request, HttpServletResponse response) {
+        // access token 확인
+        String accessToken = HeaderUtil.getAccessToken(request);
+        AuthToken authToken = tokenProvider.convertAuthToken(accessToken);
+        if (!authToken.validate()) {
+            return ApiResponse.invalidAccessToken();
+        }
+
+        // expired access token 인지 확인
+        Claims claims = authToken.getExpiredTokenClaims();
+        if (claims == null) {
+            return ApiResponse.notExpiredTokenYet();
+        }
+
+        String userId = claims.getSubject();
+        RoleType roleType = RoleType.of(claims.get("role", String.class));
+
+        // refresh token
+        String refreshToken = CookieUtil.getCookie(request, REFRESH_TOKEN)
+                .map(Cookie::getValue)
+                .orElse((null));
+        AuthToken authRefreshToken = tokenProvider.convertAuthToken(refreshToken);
+
+        if (authRefreshToken.validate()) {
+            return ApiResponse.invalidRefreshToken();
+        }
+
+        // userId refresh token 으로 DB 확인
+        UserRefreshToken userRefreshToken = userRefreshTokenRepository.findByUserIdAndRefreshToken(userId, refreshToken);
+        if (userRefreshToken == null) {
+            return ApiResponse.invalidRefreshToken();
+        }
+
+        Date now = new Date();
+        AuthToken newAccessToken = tokenProvider.createAuthToken(
+                userId,
+                roleType.getCode(),
+                new Date(now.getTime() + appProperties.getAuth().getTokenExpiry())
+        );
+
+        long validTime = authRefreshToken.getTokenClaims().getExpiration().getTime() - now.getTime();
+
+        // refresh 토큰 기간이 3일 이하로 남은 경우, refresh 토큰 갱신
+        if (validTime <= THREE_DAYS_MSEC) {
+            // refresh 토큰 설정
+            long refreshTokenExpiry = appProperties.getAuth().getRefreshTokenExpiry();
+
+            authRefreshToken = tokenProvider.createAuthToken(
+                    appProperties.getAuth().getTokenSecret(),
+                    new Date(now.getTime() + refreshTokenExpiry)
+            );
+
+            // DB에 refresh 토큰 업데이트
+            userRefreshToken.setRefreshToken(authRefreshToken.getToken());
+
+            int cookieMaxAge = (int) refreshTokenExpiry / 60;
+            CookieUtil.deleteCookie(request, response, REFRESH_TOKEN);
+            CookieUtil.addCookie(response, REFRESH_TOKEN, authRefreshToken.getToken(), cookieMaxAge);
+        }
+
+        return ApiResponse.success("token", newAccessToken.getToken());
+    }
+}
